add login and sign up, and protect the routes

+++ client/src/components/WorkoutDetail.jsx
@@ -1,12 +1,20 @@
 import formatDistanceToNow from "date-fns/formatDistanceToNow"
 import { useWorkoutsContext } from "../hooks/useWorkoutsContext"
+import { useAuthContext } from "../hooks/useAuthContext"
 
 function WorkoutDetail({ workout }) {
     const { dispatch } = useWorkoutsContext()
+    const { user } = useAuthContext()
 
     async function handleClick() {
+        if(!user) {
+            return
+        }
         const response = await fetch("/api/workouts/" + workout._id, {
-            method: "DELETE"
+            method: "DELETE",
+            headers: {
+                "Authorization": `Bearer ${user.token}`
+            }
         })
 
         const json = await response.json()
